bug:bug:: 修复权限问题 and 完善文档

+++ src/user/user.service.ts
@@ -1,4 +1,11 @@
-import { BadRequestException, GoneException, Injectable, UnauthorizedException } from "@nestjs/common";
+import {
+  BadRequestException,
+  ForbiddenException,
+  GoneException,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
 import { User } from "./user.schema";
 import { Model, mongo } from "mongoose";
@@ -36,7 +43,7 @@ export class UserService {
   async createOne(createUserDto: CreateUserDto): Promise<User> {
     const found = await this.userModel.findOne({ username: createUserDto.username }).exec();
     if (found) {
-      throw new BadRequestException("用户已经存在");
+      throw new ForbiddenException("用户已经存在");
     }
     createUserDto.password = await hashPassword(createUserDto.password);
     const createdUser = new this.userModel(createUserDto);
@@ -44,16 +51,19 @@ export class UserService {
   }
 
   async modifyOne(id: string, modifyUserDto: ModifyUserDto): Promise<User> {
+    if (modifyUserDto.role === Role.SuperAdmin) {
+      throw new ForbiddenException("不能修改为超级管理员");
+    }
     let objectId;
     try {
       objectId = new mongo.ObjectId(id);
     } catch (e) {
-      throw new BadRequestException("修改失败,请检查你的id是否有误");
+      throw new BadRequestException("id错误");
     }
     modifyUserDto.password = await hashPassword(modifyUserDto.password);
     const user = await this.userModel.findByIdAndUpdate(objectId, { $set: modifyUserDto }).exec();
     if (!user) {
-      throw new GoneException("用户不存在");
+      throw new NotFoundException("用户不存在");
     }
     return user;
   }
@@ -65,24 +75,27 @@ export class UserService {
     } catch {
       throw new BadRequestException("请检查你的id是否有误");
     }
-    const user = await this.userModel.findByIdAndDelete({ _id: objectId }).exec();
-    if (!user) {
-      throw new GoneException("用户已被删除");
+    const found = await this.userModel.findById(objectId).exec();
+    if (found.role === Role.SuperAdmin) {
+      throw new ForbiddenException("不能删除超级管理员");
     }
-    return user;
+    if (!found) {
+      throw new NotFoundException("资源不存在");
+    }
+    return await this.userModel.findByIdAndDelete({ _id: objectId }).exec();
   }
 
   async validate(user: VerifyUserDto) {
     let foundUser: User;
     foundUser = await this.userModel.findOne({ username: user.username }).exec();
     if (!foundUser) {
-      throw new BadRequestException("用户不存在");
+      throw new ForbiddenException("用户不存在");
     }
     if (await verifyPassword(user.password, foundUser.password)) {
       const { password, ...result } = foundUser.toObject();
       return result;
     } else {
-      throw new UnauthorizedException("账户或用户名错误");
+      throw new UnauthorizedException("用户名或密码错误");
     }
   }
 
@@ -92,7 +105,7 @@ export class UserService {
       username: user.username,
       role: u.role,
       _id: u._id,
-      access_token: this.jwtService.sign({ username: user.username, role: user.role }),
+      access_token: this.jwtService.sign(user),
     };
   }
 
@@ -106,16 +119,16 @@ export class UserService {
     }
     const user = await this.userModel.findByIdAndUpdate({ _id: objectId }, { $set: { password } }).exec();
     if (!user) {
-      throw new GoneException();
+      throw new NotFoundException();
     }
     return filterPassword(user);
   }
+
   async registry(user: RegistryUserDto) {
     const found = await this.userModel.findOne({ username: user.username }).exec();
     if (found) {
-      throw new BadRequestException("用户已存在");
+      throw new ForbiddenException("用户已存在");
     }
-
     user.role = Role.User;
     user.password = await hashPassword(user.password);
     const createdUser = new this.userModel(user);
